requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioredis==2.0.1
aiosignal==1.3.2
alembic==1.14.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
certifi==2024.12.14
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
loguru==0.7.3
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
redis==5.2.1
ruff==0.9.2
sentry-sdk==2.20.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
uvloop==0.21.0
yarl==1.18.3
